sdcard_image: Copy user.txt/uEnv.txt if it exists in /boot
Fixed a bug where the U-boot environment files are not copied to the boot partition if they
were installed in /boot by other recipes.

We should copy the uEnv.txt and user.txt always without placing a condition for the presence
of anything else in /boot. This is the original behavior of sdcard_image before the recent
wave of changes to it.

Signed-off-by: Joel A Fernandes <joelagnel@ti.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>