linux.inc: don't choke on missing defconfig in  lzop dep
When a recipe is being parsed which will be skipped due to an incompatible
machine, no local defconfig will exist for the current machine. It seems that
the fetch localpath code doesn't error in that case, so we need to check for an
IOError on the attempted open. Without this, we can hit parse errors.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>