ti-dmai_svn: correct path within svn checkout
At r642 as currently retrieved, the contents of ${S} are subdirectories dmai
and patches, so the subsequent attempt to run "make clean" in ${S} fails
because there is no Makefile there.  Append /dmai to the default ${S},
remove the now incorrect /dmai from LIC_FILES_CHKSUM, and correct the
strippath for the patches to reflect the new starting location.

Signed-off-by: Peter A. Bigot <bigotp@acm.org>
Acked-by: Roger Monk <r-monk@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>