linux-ti33x-psp 3.2: avoid gold linker for Thumb2 kernel
BeagleBone kernel is required to be built in Thumb2 mode for fast boot.
Gold linker is not yet perfect for this, giving an internal error:

| arm-linux-gnueabi-ld: internal error in thumb_branch_common, at .../binutils-cross-2.22-r6/binutils-2.22/gold/arm.cc:4149

While this fix can be put in generic linux.inc for all the kernels, it's
better to test the Gold linker out as much as possible and only disable it
for known issues, like this one (until it gets fixed, of course).

Signed-off-by: Denys Dmytriyenko <denys@ti.com>